# contract/sdk-generator/_cut/admin-web.openapi.yaml
openapi: 3.0.3
info:
  title: 上医智养堂 · 三端接口契约（调元云）
  version: api-contract-v1.0.0
  description: "三端（客户小程序 / 调理师·经络师 APP / 管理员 Web）一次冻结的接口契约。\n\n版本号规则（语义化，与 info.version 同值）：\n  - MAJOR（破坏级）：字段删除 / 改名 / 类型变更；可见性档位变更；枚举值语义变更；错误码语义变更；\n    涉可见性档位的 MAJOR 变更须触发「变更业务裁定」（不得按\"体验优化\"处理）。\n  - MINOR（兼容级）：新增可选请求字段 / 新增接口 / 新增响应字段。\n  - PATCH（文档级）：注释 / 示例 / 措辞修正，不动字段与语义。\n\n\U0001F6D1 措辞纪律：client_end=mp 可见字段不得含\"退款\"\"诊断\"；同步时间精确到日。\n"
  contact:
    name: 后端主程（技术单签）· 产品共签（可见性 / 措辞档位）
  license:
    name: Proprietary
servers:
- url: /api/v1
  description: Base Path（契约 §2.0 全局约定）
...
paths:
...
  /customers/{id}/band/telemetry:
    get:
      operationId: getBandTelemetry
      tags:
      - E-设备与手环数据
      summary: E3 手环原始数据 + 采集状态（四端按档裁剪）
      x-contract-row: E3
      x-callable-roles:
      - client
      - therapist
      - meridian
      - admin
      x-client-forbidden: false
      parameters:
      - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: 手环数据（已按矩阵裁剪）
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResultEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/BandTelemetryData'
        '403':
          $ref: '#/components/responses/VisibilityDenied'
...
  /doc-templates/uploads:
    post:
      operationId: uploadDocTemplate
      tags:
      - I-文书模板与渲染
      summary: I3 上传（multipart/form-data → source_type=upload）
      description: "上传约束：mime_type 仅允许三值（application/pdf /\napplication/vnd.openxmlformats-officedocument.wordprocessingml.document / text/markdown）；\nfile_size ≤ 10 MiB；落盘前先算 file_hash（SHA-256 hex），\n同 <tenant, doc_type> 下同 hash 重复上传 → 返回已有版本号（幂等返回 4002）。\n\n\U0001F6D1 反病毒 / 反脚本：上传文件不得作为可执行内容被服务端解析执行；\n   DOCX 不得解压执行宏（只做字节存储 + 哈希，不做宏解析）。\n"
      x-contract-row: I3
      x-callable-roles:
      - admin
      x-client-forbidden: false
      x-frontier: 占位待冻结
      x-idempotency-key: (tenant_id, doc_type, file_hash)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 上传结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultEnvelope'
        '403':
          $ref: '#/components/responses/TenantMismatch'
        '409':
          $ref: '#/components/responses/IdempotentReplay'
...
  /doc-templates/{id}/download:
    get:
      operationId: downloadDocTemplate
      tags:
      - I-文书模板与渲染
      summary: I7 下载原件（上传形态用）—— 仅超管
      x-contract-row: I7
      x-callable-roles:
      - admin
      x-client-forbidden: false
      x-frontier: 占位待冻结
      x-super-admin-only: true
      parameters:
      - $ref: '#/components/parameters/DocTemplateId'
      - name: version
        in: query
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 原件字节流
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          $ref: '#/components/responses/VisibilityDenied'
components:
  parameters:
    CustomerId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 客户 ID（服务端从 token 校验租户归属）
    DocTemplateId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    VisibilityDenied:
      description: 403 VISIBILITY_DENIED（该角色对请求字段组无可见性档位；data.denied_fields 回显被拒字段名）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResultEnvelope'
...
    TenantMismatch:
      description: 403 TENANT_MISMATCH（跨租户访问 / 租户头与 token 不符）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResultEnvelope'
...
    IdempotentReplay:
      description: 409 IDEMPOTENT_REPLAY（幂等键命中已有记录，返回首次结果）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResultEnvelope'
...
  schemas:
    ResultEnvelope:
      type: object
      required:
      - code
      - message
      - trace_id
      properties:
        code:
          type: integer
          description: 0 = 成功；非 0 见 x-error-codes
        message:
          type: string
          description: 面向开发者，不得直接渲染给客户（客户端有自己的 code → copy 映射）
        data:
          type: object
          nullable: true
          description: code != 0 时为空
        trace_id:
          type: string
          x-visible-to:
          - client
          - therapist
          - meridian
          - admin
      x-visibility-note: 信封字段对全部角色可见
...
    BandTelemetryData:
      type: object
      properties:
        data_source:
          type: string
          enum:
          - 手环
          - 未接入
          x-visible-to:
          - client
          - therapist
          - meridian
          - admin
          description: 未接入时数值字段不下发，不得显示 0 / 空图表 / 示例数据
        collected_days:
          type: integer
          x-field-group: capture_status
          x-visible-to:
          - client
          - therapist
          - meridian
          - admin
          description: 已采集天数（正向计数）
        synced_date:
          type: string
          format: date
          x-field-group: capture_status
          x-visible-to:
          - client
          - therapist
          - meridian
          - admin
          description: 客户可见同步时间精确到日（不得到分秒）
        metrics:
          type: array
          items:
            type: object
          x-field-group: raw_data
          x-visible-to:
          - client
          - therapist
          - meridian
          - admin
        gap_reason:
          type: string
          enum:
          - no_open
          - sync_failed
          - not_worn
          - compliant_removal
          - involuntary_technical
          - beyond_retention_window
          - unknown
          x-field-group: gap_reason
          x-visible-to:
          - therapist
          - meridian
          - admin
          description: 🔴 客户恒 403 / 不下发（硬约束，配置不得放开）
...